+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 
 ## [0.2.0] - 2022-02-24
+### Added
+- Introduced new constants `size.ObjectKeyValue` and `size.ObjectKeyUnit`.
+
 ### Changed
 - Renamed `size.NewSize` to `size.New`.
 - Optimized `size.unmarshalJSONObject` implementation.